+++ aws-sdk-resources/lib/aws-sdk-resources/services/s3.rb
@@ -5,7 +5,6 @@ module Aws
     require 'aws-sdk-resources/services/s3/object'
     require 'aws-sdk-resources/services/s3/object_summary'
     require 'aws-sdk-resources/services/s3/multipart_upload'
-    require 'aws-sdk-resources/services/s3/public_url'
 
     autoload :Encryption, 'aws-sdk-resources/services/s3/encryption'
     autoload :FilePart, 'aws-sdk-resources/services/s3/file_part'